% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_deme.R
\name{name_deme}
\alias{name_deme}
\title{Convert deme into a named list}
\usage{
name_deme(deme)
}
\arguments{
\item{deme}{}
}
\value{
\code{named deme}
}
\description{
Convert deme into a named list
}
\examples{
a <- load_yaml(text="time_units: generations
demes:
  - name: A
epochs:
  - start_size: 1000")
a <- name_deme(a)
}
